pyproject.toml
[project]
name = "nomic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.5",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
